+++ crates/icu/src/interface/mod.rs
@@ -18,10 +18,6 @@ pub mod prelude {
     pub use serde::{Deserialize, Serialize};
 }
 
-use crate::cdk::{
-    call::{CallFailed, CandidDecodeFailed, Error as CallError},
-    candid::Error as CandidError,
-};
 use thiserror::Error as ThisError;
 
 ///
@@ -35,40 +31,4 @@ pub enum InterfaceError {
 
     #[error("wasm hash matches")]
     WasmHashMatches,
-
-    #[error("call error: {0}")]
-    CallError(CallError),
-
-    #[error("call error: {0}")]
-    CallFailed(CallFailed),
-
-    #[error("candid error: {0}")]
-    CandidDecodeFailed(CandidDecodeFailed),
-
-    #[error("candid error: {0}")]
-    CandidError(CandidError),
-}
-
-impl From<CallError> for InterfaceError {
-    fn from(e: CallError) -> Self {
-        Self::CallError(e)
-    }
-}
-
-impl From<CallFailed> for InterfaceError {
-    fn from(e: CallFailed) -> Self {
-        Self::CallFailed(e)
-    }
-}
-
-impl From<CandidError> for InterfaceError {
-    fn from(e: CandidError) -> Self {
-        Self::CandidError(e)
-    }
-}
-
-impl From<CandidDecodeFailed> for InterfaceError {
-    fn from(e: CandidDecodeFailed) -> Self {
-        Self::CandidDecodeFailed(e)
-    }
 }